FIX: fixes type parameter

+++ src/PaymentOptions.java
@@ -5,7 +5,7 @@ public class PaymentOptions {
         this.price = price;
     }
 
-    public static double CalcCashPayment(int price) {
+    public static double CalcCashPayment(double price) {
 
         return 0.10 * price;
     }
